fix: 수정지시사항 단계 제외 오류 해결 및 학과관리 페이지로 이동
